classifier/classifierrequirements_win.txt
psycopg2-binary
loguru
tqdm
spacy
better_profanity
transformers==4.37.2
